+++ crates/hash/src/lib.rs
@@ -1,11 +1,10 @@
 use std::str::FromStr;
 
 use derive_more::{From, Into};
-use pyo3::prelude::*;
+use pyo3::{exceptions::PyValueError, prelude::*};
 use serde::{Deserialize, Serialize};
-use solana_program::hash::{
-    hash, Hash as HashOriginal, ParseHashError as ParseHashErrorOriginal, HASH_BYTES,
-};
+use solana_hash::{Hash as HashOriginal, ParseHashError as ParseHashErrorOriginal, HASH_BYTES};
+use solana_sha256_hasher::hash;
 use solders_macros::{common_methods, pyhash, richcmp_full};
 
 use solders_traits::handle_py_err;
@@ -133,7 +132,10 @@ impl Hash {
 
 impl PyFromBytesGeneral for Hash {
     fn py_from_bytes_general(raw: &[u8]) -> PyResult<Self> {
-        Ok(HashOriginal::new(raw).into())
+        let converted = raw
+            .try_into()
+            .map_err(|e| PyErr::new::<PyValueError, _>(format!("{:?}", e)))?;
+        Ok(HashOriginal::new_from_array(converted).into())
     }
 }